+++ Runtime/Scripts/Core/PositionValue.cs
@@ -5,6 +5,6 @@ namespace Dalichrome.RandomGenerator.Core
     public struct PositionValue
     {
         public int3 position;
-        public int value;
+        public float value;
     }
 }
